Split z into separate package
