+++ packages/example/src/posts/posts.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Post } from '@nestjs/common';
 import { createZodDto } from 'nestjs-zod'
-import { z } from 'nestjs-zod/z'
+import { z } from 'zod'
 
 class PostDto extends createZodDto(z.object({
   title: z.string().describe('The title of the post'),
