+++ packages/z/.eslintignore
@@ -1,0 +1,5 @@
+dist
+node_modules
+/*.js
+/*.ts
+!.eslintrc.js
