+++ packages/z/.gitignore
@@ -1,0 +1,22 @@
+.DS_Store
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+lerna-debug.log*
+node_modules
+.env
+.env.test
+.cache
+.next
+.nuxt
+.vscode/*
+.idea
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+dist
+package-lock.json
+/*.d.ts
+!z.d.ts
