+++ packages/z/.prettierrc
@@ -1,0 +1,16 @@
+{
+  "semi": false,
+  "singleQuote": true,
+  "tabWidth": 2,
+  "quoteProps": "consistent",
+  "endOfLine": "lf",
+  "importOrder": [
+    "^(child_process|crypto|events|fs|http|https|os|path)(\\/(.*))?$",
+    "<THIRD_PARTY_MODULES>",
+    "^[./]"
+  ],
+  "importOrderParserPlugins" : ["typescript", "decorators-legacy"],
+  "experimentalBabelParserPluginsList": [
+    "typescript"
+  ]
+}
